requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-cte==2.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
mpmath==1.3.0
openai==2.16.0
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
